# api/openapi/openapi.yaml
components:
  schemas:
...
    Cookie:
      additionalProperties: false
      properties:
        Domain:
          type: string
        Expires:
          format: date-time
          type: string
        HttpOnly:
          type: boolean
        MaxAge:
          format: int64
          type: integer
        Name:
          type: string
        Partitioned:
          type: boolean
        Path:
          type: string
        Quoted:
          type: boolean
        Raw:
          type: string
        RawExpires:
          type: string
        SameSite:
          format: int64
          type: integer
        Secure:
          type: boolean
        Unparsed:
          items:
            type: string
          type:
            - array
            - "null"
        Value:
          type: string
      required:
        - Name
        - Value
        - Quoted
        - Path
        - Domain
        - Expires
        - RawExpires
        - MaxAge
        - Secure
        - HttpOnly
        - SameSite
        - Partitioned
        - Raw
        - Unparsed
      type: object
    ErrorDetail:
      additionalProperties: false
      properties:
        location:
          description: Where the error occurred, e.g. 'body.items[3].tags' or 'path.thing-id'
          type: string
        message:
          description: Error message text
          type: string
        value:
          description: The value at the given location
      type: object
    ErrorModel:
      additionalProperties: false
      properties:
        $schema:
          description: A URL to the JSON Schema for this object.
          examples:
            - https://example.com/schemas/ErrorModel.json
          format: uri
          readOnly: true
          type: string
        detail:
          description: A human-readable explanation specific to this occurrence of the problem.
          examples:
            - Property foo is required but is missing.
          type: string
        errors:
          description: Optional list of individual error details
          items:
            $ref: "#/components/schemas/ErrorDetail"
          type:
            - array
            - "null"
        instance:
          description: A URI reference that identifies the specific occurrence of the problem.
          examples:
            - https://example.com/error-log/abc123
          format: uri
          type: string
        status:
          description: HTTP status code
          examples:
            - 400
          format: int64
          type: integer
        title:
          description: A short, human-readable summary of the problem type. This value should not change between occurrences of the error.
          examples:
            - Bad Request
          type: string
        type:
          default: about:blank
          description: A URI reference to human-readable documentation for the error.
          examples:
            - https://example.com/errors/example
          format: uri
          type: string
      type: object
...
    LoginResponse:
      additionalProperties: false
      properties:
        $schema:
          description: A URL to the JSON Schema for this object.
          examples:
            - https://example.com/schemas/LoginResponse.json
          format: uri
          readOnly: true
          type: string
        access_token:
          type: string
        csrf_token:
          description: Present only on the cookie (web) transport.
          type: string
        expires_in:
          description: Access-token lifetime in seconds.
          format: int64
          type: integer
        refresh_token:
          description: Present only on the body (native) transport.
          type: string
      required:
        - access_token
        - expires_in
      type: object
...
    RefreshRequest:
      additionalProperties: false
      properties:
        $schema:
          description: A URL to the JSON Schema for this object.
          examples:
            - https://example.com/schemas/RefreshRequest.json
          format: uri
          readOnly: true
          type: string
        refresh_token:
          type: string
      type: object
...
info:
  title: Vecingest API
  version: 0.1.0
openapi: 3.1.0
paths:
...
  /v1/auth/refresh:
    post:
      operationId: refresh
      parameters:
        - in: cookie
          name: vecingest_refresh
          schema:
            $ref: "#/components/schemas/Cookie"
        - in: header
          name: X-CSRF-Token
          schema:
            type: string
        - in: header
          name: Origin
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
          description: OK
          headers:
            Set-Cookie:
              schema:
                type: string
        default:
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorModel"
          description: Error
      summary: Rotate the refresh token
      tags:
        - auth
...